Add SSL support to kibana configs
Add several new variables to enable SSL/TLS support in Kibana, both for
the front-end and for connections to ElasticSearch.

Support for SSL/TLS was added in the "Basic" version of ElasticSearch &
Kibana starting with the 7.1.0 release.

It assumes certificates are generated externally, and only handles
uploading and configuring the server accordingly.